Show error if contract is not deployed while sending a TX (#130)
