+++ packages/nextjs/hooks/scaffold-eth/useScaffoldContractRead.ts
@@ -1,6 +1,6 @@
-import { useContractRead, useNetwork } from "wagmi";
+import { useContractRead } from "wagmi";
 import type { Abi } from "abitype";
-import { getDeployedContract } from "~~/components/scaffold-eth/Contract/utilsContract";
+import { useDeployedContractInfo } from "./useDeployedContractInfo";
 
 /**
  * @dev wrapper for wagmi's useContractRead hook which loads in deployed contract contract abi, address automatically
@@ -13,8 +13,7 @@ export const useScaffoldContractRead = (
   functionName: string,
   readConfig?: Parameters<typeof useContractRead>[0],
 ) => {
-  const { chain } = useNetwork();
-  const deployedContractData = getDeployedContract(chain?.id.toString(), contractName);
+  const deployedContractData = useDeployedContractInfo({ contractName });
 
   return useContractRead({
     functionName,
