+++ packages/nextjs/hooks/scaffold-eth/useScaffoldContractWrite.ts
@@ -1,8 +1,9 @@
 import { utils } from "ethers";
-import { useContractWrite, useNetwork, usePrepareContractWrite } from "wagmi";
-import { getDeployedContract, getParsedEthersError } from "~~/components/scaffold-eth/Contract/utilsContract";
+import { useContractWrite, usePrepareContractWrite } from "wagmi";
+import { getParsedEthersError } from "~~/components/scaffold-eth/Contract/utilsContract";
 import { toast } from "~~/utils/scaffold-eth";
 import { useTransactor } from "~~/hooks/scaffold-eth/useTransactor";
+import { useDeployedContractInfo } from "./useDeployedContractInfo";
 
 /**
  * @dev wrapper for wagmi's useContractWrite hook(with config prepared by usePrepareContractWrite hook) which loads in deployed contract abi and address automatically
@@ -12,8 +13,7 @@ import { useTransactor } from "~~/hooks/scaffold-eth/useTransactor";
  * @param value - value in ETH that will be sent with transaction
  */
 export const useScaffoldContractWrite = (contractName: string, functionName: string, args?: any[], value?: string) => {
-  const { chain } = useNetwork();
-  const deployedContractData = getDeployedContract(chain?.id.toString(), contractName);
+  const deployedContractData = useDeployedContractInfo({ contractName });
   const writeTx = useTransactor();
 
   const { config } = usePrepareContractWrite({
@@ -30,13 +30,12 @@ export const useScaffoldContractWrite = (contractName: string, functionName: str
 
   const sendContractWriteTx = async () => {
     if (!deployedContractData) {
-      toast.error("Target Contract is not defined");
+      toast.error("Target Contract is not deployed, did you forgot to run `yarn deploy`?");
       return;
     }
 
     if (wagmiContractWrite.writeAsync && writeTx) {
       try {
-        // If the contract is not deployed this "works". I get a "mined successfully" msg.
         await writeTx(wagmiContractWrite.writeAsync());
       } catch (e: any) {
         const message = getParsedEthersError(e);
